# specs/001-build-a-plugin/contracts/api.yaml
openapi: 3.0.3
info:
  title: Medusa Printify Plugin API
  description: API endpoints for Medusa Printify integration plugin
  version: 1.0.0
  contact:
    name: Medusa Printify Plugin
    url: https://github.com/your-org/medusa-plugin-printify
servers:
  - url: http://localhost:9000
    description: Local development server
  - url: https://your-store.medusa-commerce.com
    description: Production server
paths:
...
  /admin/printify/products:
    get:
      summary: List Printify products
      description: Retrieve paginated list of available Printify products
      tags:
        - Admin - Products
      security:
        - AdminAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: enabled
          in: query
          description: Filter by enablement status
          schema:
            type: boolean
        - name: search
          in: query
          description: Search products by title
          schema:
            type: string
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrintifyProductsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /store/products:
    get:
      summary: Get storefront products (filtered)
      description: Extended Medusa products endpoint with Printify filtering
      tags:
        - Store - Products
      parameters:
        - name: include_printify
          in: query
          description: Include only enabled Printify products
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreProductsResponse'
...
components:
  securitySchemes:
    AdminAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Medusa admin authentication token
    WebhookAuth:
      type: apiKey
      in: header
      name: X-Printify-Signature
      description: Printify webhook signature for verification
  schemas:
...
    PrintifyProductResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        printify_product_id:
          type: string
        medusa_product_id:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
        enabled:
          type: boolean
        last_sync_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PrintifyProductsResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/PrintifyProductResponse'
        count:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
...
    StoreProductsResponse:
      type: object
      properties:
        products:
          type: array
          items:
            type: object
            description: Standard Medusa product with Printify metadata
        count:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'